+++ spec/serega/plugins/batch/lib/plugin_extensions/activerecord_preloads_spec.rb
@@ -44,13 +44,13 @@ RSpec.describe Serega::SeregaPlugins::Batch do
       p_serializer = post_serializer
       groups = method(:groups)
       Class.new(base_serializer) do
-        attribute :first_name
-        attribute :last_name
-        attribute :posts, serializer: p_serializer, batch: {key: :id, loader: :sorted_posts}
-
         config.batch.define(:sorted_posts) do |ids|
           groups.call(AR::Post.where(user_id: ids).order(:text), by: :user_id)
         end
+
+        attribute :first_name
+        attribute :last_name
+        attribute :posts, serializer: p_serializer, batch: {key: :id, loader: :sorted_posts}
       end
     end
 
@@ -84,11 +84,6 @@ RSpec.describe Serega::SeregaPlugins::Batch do
   context "with batch loaded static data" do
     let(:user_serializer) do
       Class.new(base_serializer) do
-        attribute :first_name
-        attribute :last_name
-        attribute :posts_count, batch: {key: :id, loader: :posts_count}
-        attribute :comments_count, batch: {key: :id, loader: :comments_count}
-
         config.batch.define(:posts_count) do |ids|
           AR::Post.where(user_id: ids).group(:user_id).count
         end
@@ -96,6 +91,11 @@ RSpec.describe Serega::SeregaPlugins::Batch do
         config.batch.define(:comments_count) do |ids|
           AR::Comment.joins(:post).where(posts: {user_id: ids}).group(:user_id).count
         end
+
+        attribute :first_name
+        attribute :last_name
+        attribute :posts_count, batch: {key: :id, loader: :posts_count}
+        attribute :comments_count, batch: {key: :id, loader: :comments_count}
       end
     end
 
@@ -115,13 +115,13 @@ RSpec.describe Serega::SeregaPlugins::Batch do
       p_serializer = post_serializer
       groups = method(:groups)
       Class.new(base_serializer) do
-        attribute :first_name
-        attribute :last_name
-        attribute :posts, serializer: p_serializer, batch: {key: :id, loader: :sorted_posts}
-
         config.batch.define(:sorted_posts) do |ids|
           groups.call(AR::Post.where(user_id: ids).order(:text), by: :user_id)
         end
+
+        attribute :first_name
+        attribute :last_name
+        attribute :posts, serializer: p_serializer, batch: {key: :id, loader: :sorted_posts}
       end
     end
 
@@ -129,12 +129,12 @@ RSpec.describe Serega::SeregaPlugins::Batch do
       c_serializer = comment_serializer
       groups = method(:groups)
       Class.new(base_serializer) do
-        attribute :text
-        attribute :comments, serializer: c_serializer, batch: {key: :id, loader: :sorted_comments}
-
         config.batch.define(:sorted_comments) do |ids|
           groups.call(AR::Comment.where(post_id: ids).order(:text), by: :post_id)
         end
+
+        attribute :text
+        attribute :comments, serializer: c_serializer, batch: {key: :id, loader: :sorted_comments}
       end
     end