+++ spec/serega/plugins/batch/lib/validations/check_opt_batch_spec.rb
@@ -25,13 +25,13 @@ RSpec.describe Serega::SeregaPlugins::Batch::CheckOptBatch do
   end
 
   it "checks sub option :key is present" do
-    opts[:batch] = {loader: :abc}
+    opts[:batch] = {loader: proc {}}
     expect { check }.to raise_error Serega::SeregaError, "Option :key must present inside :batch option"
   end
 
   it "allows to skip sub option :key if default key specified" do
     serializer.config.batch.default_key = :id
-    opts[:batch] = {loader: :abc}
+    opts[:batch] = {loader: proc {}}
     expect { check }.not_to raise_error
   end
 
@@ -42,41 +42,41 @@ RSpec.describe Serega::SeregaPlugins::Batch::CheckOptBatch do
 
   it "checks sub options :key and :loader" do
     opts[:batch] = {key: :key_name, loader: :loader_name}
-    allow(Serega::SeregaPlugins::Batch::CheckBatchOptLoader).to receive(:call).with(:loader_name)
-    allow(Serega::SeregaPlugins::Batch::CheckBatchOptKey).to receive(:call).with(:key_name)
+    allow(Serega::SeregaPlugins::Batch::CheckBatchOptLoader).to receive(:call)
+    allow(Serega::SeregaPlugins::Batch::CheckBatchOptKey).to receive(:call)
 
     check
 
-    expect(Serega::SeregaPlugins::Batch::CheckBatchOptLoader).to have_received(:call).with(:loader_name)
+    expect(Serega::SeregaPlugins::Batch::CheckBatchOptLoader).to have_received(:call).with(:loader_name, serializer)
     expect(Serega::SeregaPlugins::Batch::CheckBatchOptKey).to have_received(:call).with(:key_name)
   end
 
   it "prohibits to use with :method opt" do
-    opts.merge!(batch: {key: :key, loader: :loader}, method: :method)
+    opts.merge!(batch: {key: :key, loader: proc {}}, method: :method)
     expect { check }
       .to raise_error Serega::SeregaError, "Option :batch can not be used together with option :method"
   end
 
   it "prohibits to use with :value opt" do
-    opts.merge!(batch: {key: :key, loader: :loader}, value: -> {})
+    opts.merge!(batch: {key: :key, loader: proc {}}, value: -> {})
     expect { check }
       .to raise_error Serega::SeregaError, "Option :batch can not be used together with option :value"
   end
 
   it "prohibits to use with :const opt" do
-    opts.merge!(batch: {key: :key, loader: :loader}, const: 1)
+    opts.merge!(batch: {key: :key, loader: proc {}}, const: 1)
     expect { check }
       .to raise_error Serega::SeregaError, "Option :batch can not be used together with option :const"
   end
 
   it "prohibits to use with :delegate opt" do
-    opts.merge!(batch: {key: :key, loader: :loader}, delegate: {to: :foo})
+    opts.merge!(batch: {key: :key, loader: proc {}}, delegate: {to: :foo})
     expect { check }
       .to raise_error Serega::SeregaError, "Option :batch can not be used together with option :delegate"
   end
 
   it "prohibits to use with block" do
-    opts[:batch] = {key: :key, loader: :loader}
+    opts[:batch] = {key: :key, loader: proc {}}
     expect { described_class.call(opts, proc {}, serializer) }
       .to raise_error Serega::SeregaError, "Option :batch can not be used together with block"
   end